requirements.txt
matplotlib==3.0.0
torch==1.3.0
torchvision==0.4.2
numpy==1.17.2
tqdm==4.38.0
scipy==1.3.1
GPUtil==1.4.0
python_igraph==0.7.1.post6
